refactor: gameTurn 입력을 검증하는 로직을 gameTurn 클래스 내부에 구현
